Initial commit
This automation will integrate a Xfinity XHK1-UE ZigBee Wireless Keypad with Alarmo HACS addon. It allows you to arm and disarm Alarmo via the keypad using the Alarmo passwords and not the Keypad password configured in ZHA. It will also sync the keypad state with Alarmo.

Kudos to https://github.com/gmlupatelli for the original code.